Rename Istio type directory
We need to use more Istio resources than only rbac.istio.io/v1alpha1.
Therefore, the generic name 'istio' cannot be used as a package directory.
