+++ pkg/client/clientset/versioned/typed/rbac.istio.io/v1alpha1/fake/fake_rbac.istio.io_client.go
@@ -3,7 +3,7 @@
 package fake
 
 import (
-	v1alpha1 "github.com/nais/naiserator/pkg/client/clientset/versioned/typed/istio/v1alpha1"
+	v1alpha1 "github.com/nais/naiserator/pkg/client/clientset/versioned/typed/rbac.istio.io/v1alpha1"
 	rest "k8s.io/client-go/rest"
 	testing "k8s.io/client-go/testing"
 )
