+++ pkg/client/clientset/versioned/typed/rbac.istio.io/v1alpha1/fake/fake_servicerolebinding.go
@@ -3,7 +3,7 @@
 package fake
 
 import (
-	v1alpha1 "github.com/nais/naiserator/pkg/apis/istio/v1alpha1"
+	v1alpha1 "github.com/nais/naiserator/pkg/apis/rbac.istio.io/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
