+++ pkg/client/informers/externalversions/generic.go
@@ -5,8 +5,8 @@ package externalversions
 import (
 	"fmt"
 
-	istiov1alpha1 "github.com/nais/naiserator/pkg/apis/istio/v1alpha1"
 	v1alpha1 "github.com/nais/naiserator/pkg/apis/naiserator/v1alpha1"
+	rbacistioiov1alpha1 "github.com/nais/naiserator/pkg/apis/rbac.istio.io/v1alpha1"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	cache "k8s.io/client-go/tools/cache"
 )
@@ -42,9 +42,9 @@ func (f *sharedInformerFactory) ForResource(resource schema.GroupVersionResource
 		return &genericInformer{resource: resource.GroupResource(), informer: f.Naiserator().V1alpha1().Applications().Informer()}, nil
 
 		// Group=rbac.istio.io, Version=v1alpha1
-	case istiov1alpha1.SchemeGroupVersion.WithResource("serviceroles"):
+	case rbacistioiov1alpha1.SchemeGroupVersion.WithResource("serviceroles"):
 		return &genericInformer{resource: resource.GroupResource(), informer: f.Rbac().V1alpha1().ServiceRoles().Informer()}, nil
-	case istiov1alpha1.SchemeGroupVersion.WithResource("servicerolebindings"):
+	case rbacistioiov1alpha1.SchemeGroupVersion.WithResource("servicerolebindings"):
 		return &genericInformer{resource: resource.GroupResource(), informer: f.Rbac().V1alpha1().ServiceRoleBindings().Informer()}, nil
 
 	}
