+++ pkg/client/informers/externalversions/rbac.istio.io/v1alpha1/interface.go
